max_cleaners_backend/backend/requirements.txt
asgiref==3.8.1
Django==4.1.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
pymongo==4.8.0
sqlparse==0.2.4
